backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
anthropic==0.28.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
transformers==4.41.2
torch==2.4.1
numpy==1.26.4
loguru==0.7.2
cachetools==5.3.3
wikipedia==1.4.0
reportlab==4.2.0
python-multipart==0.0.9
aiofiles==23.2.1